requirements.txt
fastapi
uvicorn
pyrofork
python-dateutil
TgCrypto